[INPLACE-285] Post doamin Application 모듈 분리 (#369)
